+++ common/src/main/java/earth/terrarium/prometheus/common/handlers/role/DefaultRole.java
@@ -1,7 +1,7 @@
 package earth.terrarium.prometheus.common.handlers.role;
 
 import com.teamresourceful.resourcefullib.common.color.Color;
-import com.teamresourceful.resourcefullib.common.utils.TriState;
+import earth.terrarium.prometheus.api.permissions.PermissionApi;
 import earth.terrarium.prometheus.api.roles.options.RoleOption;
 import earth.terrarium.prometheus.common.roles.CosmeticOptions;
 import earth.terrarium.prometheus.common.roles.HomeOptions;
@@ -17,11 +17,6 @@ public class DefaultRole {
 
     public static final UUID DEFAULT_ROLE = UUID.fromString("00000000-0000-0000-0000-000000000000");
 
-    private static final Map<String, TriState> PERMISSIONS = Util.make(new HashMap<>(), map -> {
-        map.put("headings.afk", TriState.TRUE);
-        map.put("headings.dnd", TriState.TRUE);
-    });
-
     private static final Map<ResourceLocation, RoleOption<?>> OPTIONS = Util.make(new HashMap<>(), map -> {
         map.put(CosmeticOptions.SERIALIZER.id(), new CosmeticOptions("@everyone", "⛏", Color.DEFAULT));
         map.put(HomeOptions.SERIALIZER.id(), new HomeOptions(5));
@@ -29,6 +24,6 @@ public class DefaultRole {
     });
 
     public static Role create() {
-        return new Role(PERMISSIONS, OPTIONS);
+        return new Role(PermissionApi.API.getDefaultPermissions(), OPTIONS);
     }
 }